Start breaking up relative encodings into smaller bits
